requirements.txt
fastapi
uvicorn
google-api-python-client
google-auth
python-dotenv
streamlit
gunicorn
openai>=1.3.7
langchain>=0.1.14
langchain-openai>=0.0.8